requirements.txt
Flask
PyMySQL
sqlalchemy
python-dotenv